Move remaining helper files into a utils dir
- Group together the file_aquisition.py and rendering.py files into a
utils directory because they are used throught the entire codebase.
- Also update all the imports to match their new location
