+++ deployer/commands/exec/debug/app_and_commands.py
@@ -10,8 +10,8 @@ import typer
 from ruamel.yaml import YAML
 
 from deployer.cli_app import exec_app
-from deployer.file_acquisition import find_absolute_path_to_cluster_file
 from deployer.infra_components.cluster import Cluster
+from deployer.utils.file_acquisition import find_absolute_path_to_cluster_file
 
 exec_debug_app = typer.Typer(pretty_exceptions_show_locals=False)
 exec_app.add_typer(exec_debug_app, name="debug")
