+++ deployer/commands/grafana/tokens.py
@@ -16,7 +16,7 @@ import requests
 import typer
 
 from deployer.cli_app import grafana_app
-from deployer.rendering import print_colour
+from deployer.utils.rendering import print_colour
 
 from .utils import (
     get_grafana_admin_password,
