+++ deployer/commands/validate/app_and_commands.py
@@ -14,9 +14,9 @@ import typer
 from ruamel.yaml import YAML
 
 from deployer.cli_app import validate_app
-from deployer.file_acquisition import find_absolute_path_to_cluster_file
 from deployer.infra_components.cluster import Cluster
-from deployer.rendering import print_colour
+from deployer.utils.file_acquisition import find_absolute_path_to_cluster_file
+from deployer.utils.rendering import print_colour
 
 yaml = YAML(typ="safe", pure=True)
 helm_charts_dir = Path(__file__).parent.parent.parent.parent.joinpath("helm-charts")
