+++ deployer/infra_components/cluster.py
@@ -5,8 +5,8 @@ import tempfile
 from contextlib import contextmanager
 from pathlib import Path
 
-from deployer.file_acquisition import get_decrypted_file, get_decrypted_files
-from deployer.rendering import print_colour
+from deployer.utils.file_acquisition import get_decrypted_file, get_decrypted_files
+from deployer.utils.rendering import print_colour
 
 from .hub import Hub
